+++ CHANGES.md
@@ -10,6 +10,7 @@
   - Results of a complete structure can be retrieved using `Validator.getValidationResultsFromStructure(Object)` or verified later using `Validator.verifyStructure(Object)`
   - Deprecation checks now run in the new early validation phase and check only for manually set values, not values created by later phases
   - new annotation `@Notify` to raise issues if a field is set or unset in the early validation phase.
+  - new module: klum-ast-bean-validation to provide Bean Validation integration (JSR380). (see [Validation](https://github.com/klum-dsl/klum-ast/wiki/Validation#Validation-levels-and-JSR380) [#395](https://github.com/klum-dsl/klum-ast/issues/395))
 
 ## Bugfixes
 - `StructureUtil.getPathOfFieldContaining()` and therefore `@Role` fields ignored fields where the value was actually a subclass of the field type.